+++ packages/manager/src/components/TabbedPanel/TabbedPanel.tsx
@@ -19,7 +19,7 @@ const useStyles = makeStyles((theme: Theme) => ({
     marginTop: theme.spacing(1),
   },
   panelBody: {
-    padding: `${theme.spacing(2)}px 0 0`,
+    padding: `${theme.spacing(2)} 0 0`,
   },
   tabsWrapper: {
     position: 'relative',