+++ packages/manager/src/components/TypeToConfirm/index.ts
@@ -1,2 +1,5 @@
-import TypeToConfirm from './TypeToConfirm';
+import TypeToConfirm, { Props as _TypeToConfirmProps } from './TypeToConfirm';
+
+// eslint-disable-next-line @typescript-eslint/no-empty-interface
+export interface TypeToConfirmProps extends _TypeToConfirmProps {}
 export default TypeToConfirm;