+++ packages/manager/src/components/core/ButtonBase.ts
@@ -1,6 +1,6 @@
 import ButtonBase, {
   ButtonBaseProps as _ButtonBaseProps,
-} from '@material-ui/core/ButtonBase';
+} from '@mui/material/ButtonBase';
 
 /* tslint:disable-next-line:no-empty-interface */
 export interface ButtonBaseProps extends _ButtonBaseProps {}