+++ packages/manager/src/components/core/Radio.ts
@@ -1,4 +1,4 @@
-import Radio, { RadioProps as _RadioProps } from '@material-ui/core/Radio';
+import Radio, { RadioProps as _RadioProps } from '@mui/material/Radio';
 
 /* tslint:disable-next-line:no-empty-interface */
 export interface RadioProps extends _RadioProps {}