+++ packages/validation/CHANGELOG.md
@@ -1,3 +1,14 @@
+
+## [2023-01-13] - v0.17.1
+
+### Fixed:
+- Node.js CommonJS Module Error
+
+## [2023-01-09] - v0.17.0
+
+### Changed:
+- Updated ESLint rules to not include Material UI
+
 ## [2022-11-01] - v0.16.0
 ### Changed:
 - Linode max label length from 32 to 64 characters